Refactor AuditList component to include new AuditListItem component
