+++ src/Frontend/src/components/audit/AuditList.vue
@@ -1,13 +1,10 @@
 <script setup lang="ts">
-import routeLinks from "@/router/routeLinks";
 import { FieldNames, useAuditStore } from "@/stores/AuditStore";
 import { storeToRefs } from "pinia";
-import Message, { MessageStatus } from "@/resources/Message";
 import { useRoute, useRouter } from "vue-router";
 import ResultsCount from "@/components/ResultsCount.vue";
-import { formatDotNetTimespan } from "@/composables/formatUtils";
 import FiltersPanel from "@/components/audit/FiltersPanel.vue";
-import MessageStatusIcon from "@/components/audit/MessageStatusIcon.vue";
+import AuditListItem from "@/components/audit/AuditListItem.vue";
 import { onBeforeMount, onUnmounted, ref, watch } from "vue";
 import RefreshConfig from "../RefreshConfig.vue";
 import useAutoRefresh from "@/composables/autoRefresh";
@@ -36,15 +33,6 @@ const dataRetriever = useAutoRefresh(
 onUnmounted(() => {
   dataRetriever.updateTimeout(null);
 });
-
-function navigateToMessage(message: Message) {
-  const query = router.currentRoute.value.query;
-
-  router.push({
-    path: message.status === MessageStatus.Successful || message.status === MessageStatus.ResolvedSuccessfully ? routeLinks.messages.successMessage.link(message.message_id, message.id) : routeLinks.messages.failedMessage.link(message.id),
-    query: { ...query, ...{ back: route.path } },
-  });
-}
 
 const firstLoad = ref(true);
 
@@ -125,17 +113,7 @@ watch(autoRefreshValue, (newValue) => dataRetriever.updateTimeout(newValue));
     <div class="row results-table">
       <LoadingSpinner v-if="firstLoad" />
       <template v-for="message in messages" :key="message.id">
-        <div class="item" @click="navigateToMessage(message)">
-          <div class="status">
-            <MessageStatusIcon :message="message" />
-          </div>
-          <div class="message-id">{{ message.message_id }}</div>
-          <div class="message-type">{{ message.message_type }}</div>
-          <div class="time-sent"><span class="label-name">Time Sent:</span>{{ new Date(message.time_sent).toLocaleString() }}</div>
-          <div class="critical-time"><span class="label-name">Critical Time:</span>{{ formatDotNetTimespan(message.critical_time) }}</div>
-          <div class="processing-time"><span class="label-name">Processing Time:</span>{{ formatDotNetTimespan(message.processing_time) }}</div>
-          <div class="delivery-time"><span class="label-name">Delivery Time:</span>{{ formatDotNetTimespan(message.delivery_time) }}</div>
-        </div>
+        <AuditListItem :message="message" />
       </template>
     </div>
   </div>
@@ -158,52 +136,5 @@ watch(autoRefreshValue, (newValue) => dataRetriever.updateTimeout(newValue));
   margin-top: 1rem;
   margin-bottom: 5rem;
   background-color: #ffffff;
-}
-.item {
-  padding: 0.3rem 0.2rem;
-  border: 1px solid #ffffff;
-  border-bottom: 1px solid #eee;
-  display: grid;
-  grid-template-columns: 1.8em 1fr 1fr 1fr 1fr;
-  grid-template-rows: 1fr 1fr;
-  gap: 0.375rem;
-  grid-template-areas:
-    "status message-type message-type message-type time-sent"
-    "status message-id processing-time critical-time delivery-time";
-}
-.item:not(:first-child) {
-  border-top-color: #eee;
-}
-.item:hover {
-  border-color: #00a3c4;
-  background-color: #edf6f7;
-  cursor: pointer;
-}
-.label-name {
-  margin-right: 0.25rem;
-  color: #777f7f;
-}
-.status {
-  grid-area: status;
-}
-.message-id {
-  grid-area: message-id;
-}
-.time-sent {
-  grid-area: time-sent;
-}
-.message-type {
-  grid-area: message-type;
-  font-weight: bold;
-  overflow-wrap: break-word;
-}
-.processing-time {
-  grid-area: processing-time;
-}
-.critical-time {
-  grid-area: critical-time;
-}
-.delivery-time {
-  grid-area: delivery-time;
 }
 </style>
